udpate

+++ core/src/com/unciv/ui/mapeditor/MapEditorSaveTab.kt
@@ -9,13 +9,14 @@ import com.unciv.logic.MapSaver
 import com.unciv.logic.map.MapType
 import com.unciv.logic.map.TileMap
 import com.unciv.models.translations.tr
+import com.unciv.ui.popup.ConfirmPopup
 import com.unciv.ui.popup.Popup
 import com.unciv.ui.popup.ToastPopup
-import com.unciv.ui.popup.YesNoPopup
 import com.unciv.ui.utils.AutoScrollPane
 import com.unciv.ui.utils.BaseScreen
 import com.unciv.ui.utils.KeyCharAndCode
 import com.unciv.ui.utils.TabbedPager
+import com.unciv.ui.utils.UncivTextField
 import com.unciv.ui.utils.extensions.isEnabled
 import com.unciv.ui.utils.extensions.keyShortcuts
 import com.unciv.ui.utils.extensions.onActivation
@@ -37,7 +38,7 @@ class MapEditorSaveTab(
     private val deleteButton = "Delete map".toTextButton()
     private val quitButton = "Exit map editor".toTextButton()
 
-    private val mapNameTextField = TextField("", skin)
+    private val mapNameTextField = UncivTextField.create("Map Name")
 
     private var chosenMap: FileHandle? = null
 
@@ -81,7 +82,11 @@ class MapEditorSaveTab(
 
     private fun deleteHandler() {
         if (chosenMap == null) return
-        YesNoPopup("Are you sure you want to delete this map?", editorScreen) {
+        ConfirmPopup(
+            editorScreen,
+            "Are you sure you want to delete this map?",
+            "Delete map",
+        ) {
             chosenMap!!.delete()
             mapFiles.update()
         }.open()
